fix: fix run ci error
